Split the transport-agnostic half of the boot menu out of IpxeBootMenu
getTasking() decides what a tasked host boots and with which kernel
arguments. None of that reasoning is iPXE's -- it is FOG's -- but it has
always lived inside IpxeBootMenu because iPXE was the only consumer.
A second bootloader needs the identical decisions and a completely
different output syntax, so this moves the decisions into an abstract
BootMenuBase and leaves each renderer to say them in its own language.

Moved verbatim: getTasking(), _wantsShutdown(), _extraArgs(), and the
$_kernel/$_storage properties they read. Four hooks stay with the
renderer -- printDefault(), _printImageIgnored(), _printTasking() and a
new _printMemtestTasking(), which is the one place the split is not a
straight move: the memtest branch built an iPXE $Send array inline, and
memtest is a memdisk image rather than a kernel-and-init chain, so the
hook takes no argument set at all.

Nothing about iPXE's output changes, and the golden file is the proof:
the 46 scenarios committed in the previous commit render byte-identical
across the move.

The 47th scenario is new and closes a hole that commit left. No
scenario asked for a shutdown, so _wantsShutdown() and _extraArgs()
were being moved with zero coverage -- and _wantsShutdown() is the
helper that exists because stripos()'s arguments were reversed at five
call sites. task-deploy-shutdown asks for one via extraargs alone, with
the task type's own kernelArgs deliberately silent on it, so the render
can only emit shutdown=1 by reading both helpers correctly. Proven by
mutation: returning false, returning an empty extraargs, and restoring
the original reversed stripos() each turn the test red.

bin/psr4-scan.php gains the taxonomy entry --check refuses to guess,
and the render harness requires the base before the class under test
because it has no autoloader.
